requirements.txt
# Telegram Bot Framework - Latest 2025 Versions
python-telegram-bot==22.3

# Database (let motor resolve compatible pymongo)
motor==3.6.0

# Web Scraping & HTTP Requests
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.46.0

# Async Support
asyncio-mqtt==0.16.2
aiohttp==3.9.5

# Task Scheduling
APScheduler==3.10.4

# Environment Variables
python-dotenv==1.0.1

# Data Processing
# Use environment markers to select wheels per Python version
# Py < 3.13 uses stable 1.26.x/2.2.2 wheels; Py >= 3.13 needs newer wheels
numpy==1.26.4; python_version < "3.13"
numpy>=2.1,<2.3; python_version >= "3.13"
pandas==2.2.2; python_version < "3.13"
pandas>=2.2.3,<2.4; python_version >= "3.13"

# Logging & Monitoring
structlog==24.4.0

# Date & Time
python-dateutil==2.9.0

# Web Server for Health Check (Render requirement)
uvicorn==0.30.5
fastapi==0.112.0

# Optional: Enhanced async performance (skip on Python 3.13+ to avoid source build)
uvloop==0.19.0; platform_system != "Windows" and python_version < "3.13"

# Greenlet wheels compatibility with newer Python (Python 3.13 requires >=3.1.0)
greenlet>=3.1.0
